Move elkm1 services to separate module
